+++ packages/react-ui/storybook/stories/widgetsUI/HistogramWidgetUI.stories.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import HistogramWidgetUI from '../../../src/widgets/HistogramWidgetUI';
-import { buildReactPropsAsString } from '../../utils';
 
 const options = {
   title: 'Widgets UI/HistogramWidgetUI',
@@ -24,7 +23,6 @@ const EmptyProps = {
   tooltipFormatter: () => {}
 };
 Empty.args = EmptyProps;
-Empty.parameters = buildReactPropsAsString(EmptyProps, 'HistogramWidgetUI');
 
 export const Simple = Template.bind({});
 const SimpleProps = {
@@ -36,7 +34,6 @@ const SimpleProps = {
   tooltipFormatter: (params) => params[0].value + ' $'
 };
 Simple.args = SimpleProps;
-Simple.parameters = buildReactPropsAsString(SimpleProps, 'HistogramWidgetUI');
 
 export const xAxisFormatter = Template.bind({});
 const xAxisFormatterProps = {
@@ -47,10 +44,6 @@ const xAxisFormatterProps = {
   tooltipFormatter: (params) => params[0].value + ' $'
 };
 xAxisFormatter.args = xAxisFormatterProps;
-xAxisFormatter.parameters = buildReactPropsAsString(
-  xAxisFormatterProps,
-  'HistogramWidgetUI'
-);
 
 export const yAxisFormatter = Template.bind({});
 const yAxisFormatterProps = {
@@ -61,10 +54,6 @@ const yAxisFormatterProps = {
   tooltipFormatter: (params) => params[0].value + ' $'
 };
 yAxisFormatter.args = yAxisFormatterProps;
-yAxisFormatter.parameters = buildReactPropsAsString(
-  yAxisFormatterProps,
-  'HistogramWidgetUI'
-);
 
 export const Filtered = Template.bind({});
 const FilteredProps = {
@@ -76,4 +65,3 @@ const FilteredProps = {
   onSelectedBarsChange: (evt) => {}
 };
 Filtered.args = FilteredProps;
-Filtered.parameters = buildReactPropsAsString(FilteredProps, 'HistogramWidgetUI');