+++ packages/react-ui/webpack.config.js
@@ -6,7 +6,7 @@ const webpackBaseConfig = require('../../webpack.base');
 
 const webpackConfig = {
   ...webpackBaseConfig,
-  externals: [ 
+  externals: [
     /^@material-ui\/.+$/,
     '@carto/react-core',
     'echarts',