+++ packages/react-core/src/operations/GroupDateTypes.d.ts
@@ -3,6 +3,6 @@ export enum GroupDateTypes {
   MONTHS = 'months',
   WEEKS = 'weeks',
   DAYS = 'days',
+  HOURS = 'hours',
+  MINUTES = 'minutes'
 }
-
-